requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
crispy-bootstrap5==2024.10
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-allauth==65.3.0
django-countries==7.6.1
django-crispy-forms==2.3
django-decouple==2.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
idna==3.10
jwt==1.3.1
oauthlib==3.2.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
